Remove CTREModuleState
* Hated having that class, it was kind of extraneous and the structure did not entirely work.
* Implemented custom solution. :+1:
* Moved expand module state method to SwerveModlue class.
* Increased jitter threshold by 2%
